Multibody decay tests fixed
The "default" if no wave options given is to have no waves, there is a bug with this for multibody systems atm. For now, the No Wave option is explicitly given in multibody cases.
TODO: fix default no wave option for multibody, issue to be added to github soon
Also, continued adjusting file paths for some output in demos